% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\docType{data}
\name{colr}
\alias{colr}
\title{Colr}
\format{An object of class \code{data.frame} with 10 rows and 3 columns.}
\usage{
colr
}
\description{
Data frame of AU colors in wide format. Light (standard) and dark in separate columns.
}
\keyword{datasets}
