% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\docType{data}
\name{colrlong}
\alias{colrlong}
\title{Long data frame of AU colors}
\format{An object of class \code{data.frame} with 20 rows and 2 columns.}
\usage{
colrlong
}
\description{
Data frame of AU colors in long format. Light (standard) and dark in same column.
}
\keyword{datasets}
