% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\name{paletdark}
\alias{paletdark}
\title{Palete dark}
\usage{
paletdark()
}
\value{
A vector of dark color hex codes.
}
\description{
Palette of all dark colors, printet as character vector. Does not take any arguments.
}
