% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\name{pallet}
\alias{pallet}
\title{Pallete}
\usage{
pallet()
}
\value{
A vector of light color hex codes.
}
\description{
Pallette of all light colors, printet as character vector. Does not take any arguments.
}
