% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\name{palletdark}
\alias{palletdark}
\title{Pallete dark}
\usage{
palletdark()
}
\value{
A vector of dark color hex codes.
}
\description{
Pallette of all dark colors, printet as character vector. Does not take any arguments.
}
