% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\name{palletlong}
\alias{palletlong}
\title{Pallete long}
\usage{
palletlong()
}
\value{
A vector of all color hex codes.
}
\description{
Pallette of all colors, printet as character vector. Does not take any arguments.
}
