% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\name{picker}
\alias{picker}
\title{Color picker}
\usage{
picker(x = c("blue"))
}
\arguments{
\item{x}{A vector of color names as specified on the AU colors webpage. Dark versions a names "dark [color]". Only takes lower case names.}
}
\value{
A vector of hex codes for the specified colors
}
\description{
Color picker, takes a vector of colornames and returns a hex codes.
Defaults to AU blue.
All color names need to be lower case
}
