% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucolr.R
\name{paletlong}
\alias{paletlong}
\title{AU color palette, standard followed by dark}
\usage{
paletlong()
}
\value{
A vector of all color hex codes.
}
\description{
palette of all colors, printet as character vector. Does not take any arguments.
}
